Next Step towards a Unified, Automated Property Value Grouping
Just the next step, code is not working yet. This is another
intermediate commit. Hopefully I get it working tomorrow.

+++ src/main/java/org/optimizationBenchmarking/experimentation/evaluation/attributes/clusters/propertyValueGroups/EGroupingMode.java
@@ -368,7 +368,7 @@ public enum EGroupingMode {
     return new _Groups(buffer, minGroups, maxGroups,
         ((power == 2L) ? POWERS_OF_2 : //
             ((power == 10L) ? POWERS_OF_10 : ANY_POWERS)),//
-        Double.valueOf(power));
+        Long.valueOf(power));
   }
 
   /**
